refactor(checker): split call elaboration helpers
